Tweak example to pass in newer GAP versions, too
On GAP master there recently was an improvement made to MinimalGeneratingSet
for non-pc groups. It turns out that this affects an example in the UGALY
manual which is also exercised by the test suite. This results in a different
output for `InvolutiveCompatibilityCocycle` and subsequently also for
`LocalActionGamma`, resulting in the test failure

This patch tweaks the example to only test the permutation type of the group
returned by `LocalActionGamma`.